HEEDLS-621 Move view model to the ViewModel folder
