+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/DelegateProgress/EditCompletionDate.cshtml
@@ -1,6 +1,6 @@
 @using DigitalLearningSolutions.Web.Models.Enums
 @using DigitalLearningSolutions.Web.ViewComponents
-@using DigitalLearningSolutions.Web.Views.TrackingSystem.Delegates.DelegateProgress
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateProgress
 @using Microsoft.AspNetCore.Mvc.TagHelpers
 @model EditCompletionDateViewModel
