--- DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/DelegateProgress/EditCompletionDateViewModel.cs
@@ -1,36 +1,0 @@
-// QQ fix line endings before merge
-
-namespace DigitalLearningSolutions.Web.Views.TrackingSystem.Delegates.DelegateProgress
-{
-    using DigitalLearningSolutions.Data.Models.Courses;
-    using DigitalLearningSolutions.Web.Models.Enums;
-    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateProgress;
-
-    public class EditCompletionDateViewModel : EditCompletionDateFormData
-    {
-        public EditCompletionDateViewModel(
-            int progressId,
-            DelegateProgressAccessRoute accessedVia,
-            DelegateCourseInfo info,
-            int? returnPage
-        ) : base(info, returnPage)
-        {
-            ProgressId = progressId;
-            AccessedVia = accessedVia;
-        }
-
-        public EditCompletionDateViewModel(
-            EditCompletionDateFormData formData,
-            int progressId,
-            DelegateProgressAccessRoute accessedVia
-        ) : base(formData)
-        {
-            ProgressId = progressId;
-            AccessedVia = accessedVia;
-        }
-
-        public int ProgressId { get; set; }
-
-        public DelegateProgressAccessRoute AccessedVia { get; set; }
-    }
-}
